extract LazyParseTransforms
